General Bugfix (Performance + wrong behaviour)
* Fixes some crashes
* Previous performance issues due to tiles[][] arrays being humongus
  -> Wrong allocation size
* Patches some wrong behaviour, f.e. in culling

+++ src/com/prototype3/gameobjects/PhysicsObject.java
@@ -8,8 +8,8 @@ public class PhysicsObject extends GameObject {
 	public int y;
 	public int width;
 	public int height;
-	public int speedX;
-	public int speedY;
+	public float speedX;
+	public float speedY;
 	public boolean isStatic;
 
 	public PhysicsObject(int x, int y, int width, int height) {
